requirements.txt
requests
selenium
python-dotenv
streamlit
PyPDF2
python-docx
beautifulsoup4
webdriver-manager